+++ src/main/java/com/ufes/delivery/model/state/ReembolsadoState.java
@@ -13,7 +13,7 @@ public class ReembolsadoState extends PedidoState {
     }
 
     @Override
-    public void avaliar(double nota) {
+    public void avaliar(int nota) {
         super.avaliar(nota);
     }
